Quickfix

+++ code/modules/mining/equipment_locker.dm
@@ -361,7 +361,7 @@
 	var/datum/effect/effect/system/spark_spread/s = new /datum/effect/effect/system/spark_spread
 	s.set_up(5, 1, src)
 	s.start()
-	if(prob(50 / severity) || severity < 3)
+	if(prob(50 / severity) && severity < 3)
 		qdel(src)
 
 /**********************Mining Equipment Vendor Items**************************/
